Summary
Added several images that might become my background image.  Trying to make the background image fixed and sized to fit the page.  The picture is being blocked by the jumbotron section.  I will try to let the image go over that.  I will also include a fixed table on the right that will contain more sections of the fan site for more info on the characters, seasons, episodes, and maps to help give insight to the series.